Bug 347 and 423: - Add an action to delete unreferenced definitions - Create a new preference page to keep preferences about references and definitions - Support deleting the definitions when deleting the last reference. (add a message box to prompt the user/base on the preferences) - Support deleting all the references when deleting the a definition (add a message box to prompt the user/base on the preferences)
git-svn-id: svn://cserg0.site.uottawa.ca/projetseg/trunk@1546 1d5a7336-c65e-6947-a78c-d4236001c669

+++ src/seg/jUCMNav/model/commands/delete/DeleteActorCommand.java
@@ -4,13 +4,17 @@
 package seg.jUCMNav.model.commands.delete;
 
 import grl.Actor;
+import grl.ActorRef;
 
 import java.util.Iterator;
 
 import org.eclipse.gef.commands.CompoundCommand;
 
 import seg.jUCMNav.Messages;
+import seg.jUCMNav.model.commands.delete.internal.PreDeleteUrnModelElementCommand;
 import seg.jUCMNav.model.commands.delete.internal.RemoveActorCommand;
+import seg.jUCMNav.model.commands.delete.internal.RemoveURNmodelElementCommand;
+import seg.jUCMNav.views.preferences.DeletePreferences;
 import urn.URNlink;
 
 /**
@@ -21,23 +25,74 @@ import urn.URNlink;
  */
 public class DeleteActorCommand extends CompoundCommand{
 
+	private Actor actor;
+	
     /**
      * @param actor
      *          the ActorRef to delete
      */
-    public DeleteActorCommand(Actor actor) {
+	public DeleteActorCommand(Actor actor) {
         setLabel(Messages.getString("DeleteActorCommand.deleteActor")); //$NON-NLS-1$
-        
-        //Remove the URNlinks
-        for (Iterator it = actor.getFromLinks().iterator(); it.hasNext();){
-            URNlink link = (URNlink)it.next();
-            add(new DeleteURNlinkCommand(link));
+        this.actor = actor;
+    }
+
+    /**
+     * Returns true even if no commands exist.
+     */
+    public boolean canExecute() {
+        if (getCommands().size() == 0)
+            return true;
+        else
+            return super.canExecute();
+    }
+    
+    /**
+     * Returns true even if no commands exist.
+     */
+    public boolean canUndo() {
+        if (getCommands().size() == 0)
+            return true;
+        else
+            return super.canUndo();
+    }
+    
+    /**
+     * Late building
+     */
+    public void execute() {
+        build();
+        super.execute();
+    }
+    
+    /**
+     * Build the compound command.
+     * 
+     */
+    private void build()
+    {
+        //Verify if the definition can be delete.
+        if(actor.getContRefs().size() == 0 ||
+        		DeletePreferences.getDeleteReference(actor))
+        {
+	        //Remove all the actor references
+	        for(Iterator it = actor.getContRefs().iterator(); it.hasNext(); )
+	        {
+	        	ActorRef actorRef = (ActorRef)it.next();
+	            add(new PreDeleteUrnModelElementCommand(actorRef));
+	            add(new RemoveURNmodelElementCommand(actorRef));
+	        }
+	        
+	        //Remove the URNlinks
+	        for (Iterator it = actor.getFromLinks().iterator(); it.hasNext();){
+	            URNlink link = (URNlink)it.next();
+	            add(new DeleteURNlinkCommand(link));
+	        }
+	        for (Iterator it = actor.getToLinks().iterator(); it.hasNext();){
+	            URNlink link = (URNlink)it.next();
+	            add(new DeleteURNlinkCommand(link));
+	        }
+	        add(new RemoveActorCommand(actor));
         }
-        for (Iterator it = actor.getToLinks().iterator(); it.hasNext();){
-            URNlink link = (URNlink)it.next();
-            add(new DeleteURNlinkCommand(link));
-        }
-        add(new RemoveActorCommand(actor));
     }
 
 }
