+++ src/seg/jUCMNav/model/commands/delete/DeleteIntentionalElementCommand.java
@@ -7,6 +7,7 @@ import grl.ElementLink;
 import grl.Evaluation;
 import grl.EvaluationStrategy;
 import grl.IntentionalElement;
+import grl.IntentionalElementRef;
 import grl.kpimodel.Indicator;
 import grl.kpimodel.KPIModelLink;
 
@@ -15,7 +16,10 @@ import java.util.Iterator;
 import org.eclipse.gef.commands.CompoundCommand;
 
 import seg.jUCMNav.Messages;
+import seg.jUCMNav.model.commands.delete.internal.PreDeleteUrnModelElementCommand;
 import seg.jUCMNav.model.commands.delete.internal.RemoveIntentionalElementCommand;
+import seg.jUCMNav.model.commands.delete.internal.RemoveURNmodelElementCommand;
+import seg.jUCMNav.views.preferences.DeletePreferences;
 import urn.URNlink;
 
 /**
@@ -49,6 +53,16 @@ public class DeleteIntentionalElementCommand extends CompoundCommand {
         return super.canExecute();
     }
 
+    /**
+     * Returns true even if no commands exist.
+     */
+    public boolean canUndo() {
+        if (getCommands().size() == 0)
+            return true;
+        else
+            return super.canUndo();
+    }
+    
     private void deleteElementLink() {
         if (element.getLinksDest().size() > 0 || element.getLinksSrc().size() > 0) {
             for (int i = 0; i < element.getLinksDest().size(); i++) {
@@ -78,7 +92,7 @@ public class DeleteIntentionalElementCommand extends CompoundCommand {
     // TODO Fix delete Evaluations.
     // A modification is required in the metamodel to implement a bidirectional links between Intentional Elements and Evaluations (instead of unidirectional).
     // Should access the evaluation to delete using this association.
-    private void deleteEvaluations() {
+	private void deleteEvaluations() {
         for (Iterator it = element.getGrlspec().getStrategies().iterator(); it.hasNext();) {
             EvaluationStrategy strategy = (EvaluationStrategy) it.next();
             for (Iterator itEval = strategy.getEvaluations().iterator(); itEval.hasNext();) {
@@ -94,7 +108,7 @@ public class DeleteIntentionalElementCommand extends CompoundCommand {
      * Late building
      */
     public void execute() {
-        build();
+    	build();
         super.execute();
     }
 
@@ -102,24 +116,39 @@ public class DeleteIntentionalElementCommand extends CompoundCommand {
      * Builds a sequence of DeleteGRLNodeCommands
      * 
      */
-    private void build() {
-        // Delete all the URNlink
-        for (Iterator it = element.getFromLinks().iterator(); it.hasNext();) {
-            URNlink link = (URNlink) it.next();
-            add(new DeleteURNlinkCommand(link));
+	private void build() {
+
+        //Verify if the definition can be delete.
+        if(element.getRefs().size() == 0 ||
+        		DeletePreferences.getDeleteReference(element))
+        {
+	    	// Delete all the URNlink
+	        for (Iterator it = element.getFromLinks().iterator(); it.hasNext();) {
+	            URNlink link = (URNlink) it.next();
+	            add(new DeleteURNlinkCommand(link));
+	        }
+	        for (Iterator it = element.getToLinks().iterator(); it.hasNext();) {
+	            URNlink link = (URNlink) it.next();
+	            add(new DeleteURNlinkCommand(link));
+	        }
+	        deleteEvaluations();
+	
+	        // Delete all the ElementLink associate with the IntentionalElement
+	        deleteElementLink();
+	
+	        // Delete all the KPIModelLink associate with the IntentionalElement
+	        deleteKPIModelLink();
+	
+	        //Delete of all the references
+	    	for(Iterator it=element.getRefs().iterator(); it.hasNext(); )
+	    	{
+	    		IntentionalElementRef reference = (IntentionalElementRef)it.next();
+	            add(new PreDeleteUrnModelElementCommand(reference));
+	            add(new RemoveURNmodelElementCommand(reference));
+	    	}
+	
+	        add(new RemoveIntentionalElementCommand(element));
         }
-        for (Iterator it = element.getToLinks().iterator(); it.hasNext();) {
-            URNlink link = (URNlink) it.next();
-            add(new DeleteURNlinkCommand(link));
-        }
-        deleteEvaluations();
 
-        // Delete all the ElementLink associate with the IntentionalElement
-        deleteElementLink();
-
-        // Delete all the KPIModelLink associate with the IntentionalElement
-        deleteKPIModelLink();
-
-        add(new RemoveIntentionalElementCommand(element));
     }
 }
